Fixes
Hi, Arjin,

There are a number of fixes that I made on this, but here are the primary things that I did.

The `"noteon"`and `"noteoff"` listeners were calling the `MIDIprocess` function instead of `playChord`. 

`playChord` wasn't returning anything, so I added a `return` statement. Additionally, I had push all of the notes in the chord to an array, and then return that array.


I also created a `chords` object to store the different chord intervals, and had that dynamically populate the dropdown menu.

Happy to go over this in more detail later, but thanks for getting this in.